Introduce functional datachain API (#991)
* Introduce functional datachain API

- Updated import statements to use `import datachain as dc` for consistency across examples and documentation.
- Refactored example code snippets in `examples.md` and `quick-start.md` to reflect the new import style.
- Enhanced documentation in `datachain.md`, `file.md`, `imagefile.md`, `textfile.md`, and `videofile.md` to clarify usage of `from_storage` and other methods.
- Introduced new utility functions in `dc` module for CSV, JSON, and other data formats to streamline data ingestion processes.
- Added new methods for handling datasets and listings, improving the overall functionality of the DataChain library.

Closes #984

* Fix e2e
