+++ tests/scripts/feature_class.py
@@ -1,6 +1,6 @@
 from pydantic import BaseModel
 
-from datachain.lib.dc import C, DataChain
+import datachain as dc
 
 
 class Embedding(BaseModel):
@@ -9,8 +9,8 @@ class Embedding(BaseModel):
 
 ds_name = "feature_class"
 ds = (
-    DataChain.from_storage("gs://dvcx-datalakes/dogs-and-cats/")
-    .filter(C("file.path").glob("*cat*.jpg"))
+    dc.from_storage("gs://dvcx-datalakes/dogs-and-cats/")
+    .filter(dc.C("file.path").glob("*cat*.jpg"))
     .order_by("file.path")
     .limit(5)
     .map(emd=lambda file: Embedding(value=512), output=Embedding)
